Multiple Tickers per string & parses post body
Any number of tickers can now be parsed in a given string, instead of just the first appearing ticker. Additionally, the body of the post is parsed for tickers, not just the title (some DD posts, especially on r/pennystocks, list DD for multiple different companies in the body of the post).